Move BlendModeExtensions into ImageEdit
